requirements.txt
# Topic Analysis Pipeline Dependencies

# Data Processing
polars>=0.20.0
numpy>=1.24.0

# Vector Database
chromadb>=0.4.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.0
hdbscan>=0.8.0
scikit-learn>=1.3.0

# Statistical Analysis
statsmodels>=0.14.0
scipy>=1.11.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
streamlit>=1.30.0
pandas>=2.0.0

# AI/LLM
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
